feat: Support automatically finding sensor data. (#27)
The scope of changes include modifications to field values in the
config.yml file.  These changes include the name of the hwmon component
(i.e. coretemp for the cpu, amdgpu for the gpu, etc.) and a label name
for the sensor to read (i.e. package for cpu and edge for gpu).

During the call to New for the CPU and GPU components, an attempt will
be made to find the full system path to the desired sensor data.  If not
found, an error will be returned.